--- dmm/item_search_test.go
@@ -1,9 +1,0 @@
-package dmm
-
-import (
-	"testing"
-)
-
-func TestNewItemSearchClientImpl(t *testing.T) {
-	NewItemSearchClientImpl("dmm api id", "dmm affiliate id")
-}
