+++ domain/repository/item_search_repository.go
@@ -1,0 +1,7 @@
+package repository
+
+import "github.com/otknoy/dmm-crawler/domain/model"
+
+type ItemSearchRepository interface {
+	Search(keyword string, hits int, offset int) (model.ItemResponse, error)
+}
